All classes are compatible and are able to be built upon
