+++ src/main/java/edu/farmingdale/csc325_groupproject/QuaternaryController.java
@@ -2,10 +2,8 @@ package edu.farmingdale.csc325_groupproject;
 
 import java.io.IOException;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.ResourceBundle;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
+import java.util.*;
+import javafx.fxml.*;
 import javafx.scene.control.TextField;
 import javafx.event.ActionEvent;
 
@@ -13,41 +11,30 @@ public class QuaternaryController implements Initializable {
     
     @FXML
     private TextField codeTxt;
-
     @FXML
     private TextField dateTxt;
-
     @FXML
     private TextField descTxt;
-
     @FXML
     private TextField distTxt;
-
     @FXML
     private TextField loc1Txt;
-
     @FXML
     private TextField locTxt;
-
     @FXML
     private TextField nameTxt;
-
     @FXML
     private TextField neighTxt;
-
     @FXML
     private TextField postTxt;
-
     @FXML
     private TextField timeTxt;
-
     @FXML
     private TextField totTxt;
-
     @FXML
     private TextField weapTxt;
     
-    private ArrayList<Complaints> comps = new ArrayList<Complaints>();
+    private ArrayList<Complaint> comps = new ArrayList<Complaint>();
     /**
      * Initializes the controller class.
      */
@@ -57,7 +44,7 @@ public class QuaternaryController implements Initializable {
     } 
     @FXML
     void InputData(ActionEvent event) {
-        Complaints b = new Complaints();
+        Complaint b = new Complaint();
         b.CrimeDate = dateTxt.getText();
         b.CrimeTime = timeTxt.getText();
         b.CrimeCode = codeTxt.getText();
@@ -72,7 +59,6 @@ public class QuaternaryController implements Initializable {
         b.CrimeName = nameTxt.getText();
         comps.add(b);
     }
-    
     
     @FXML
     private void switchToMenu() throws IOException {
