+++ src/main/java/edu/farmingdale/csc325_groupproject/TertiaryController.java
@@ -3,11 +3,10 @@ package edu.farmingdale.csc325_groupproject;
 import java.io.IOException;
 import java.net.URL;
 import java.sql.*;
+import javafx.fxml.*;
 import java.util.ResourceBundle;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
 import javafx.scene.control.*;
 
 public class TertiaryController implements Initializable {
